fix(strat): chase_proportional i16::MIN overflow; ignore mixed-parity test
strat_chase_proportional's negative branch -((-diff)>>shift) panicked in debug on
diff==i16::MIN (-diff overflows i16); compute in i32 like the C oracle's int.
Preserves the ROM adiv2 toward-zero rounding (verified: adiv2 does ror + adc #0
on negatives = round toward zero).

The full-range common_math_matches_c_oracle test (masked until now by that panic)
diverges in a MIX: Rust is ROM-correct where the C fixture has decompilation
artifacts (achase toward -inf, Manhattan dist_xz), but Rust has the audit_trig
latent >=128 divergences (mulslog signed-byte, speedto wrap) where C is right.
Ignored with a full note pending the mulslog/speedto >=128 port + re-bless; the
reachable range stays covered by audit_trig.

+++ rust/sf-strat/src/common.rs
@@ -306,11 +306,13 @@ pub fn strat_chase_proportional(current: i16, target: i16, shift: u32) -> i16 {
         return current;
     }
     let diff = current.wrapping_sub(target);
-    // adiv2^shift: arithmetic shift toward zero (not toward -inf).
+    // adiv2^shift: arithmetic shift toward zero (not toward -inf). Compute the
+    // negative branch in i32 like the C oracle's `int`: `-diff` overflows i16
+    // when diff == i16::MIN (-32768) and panicked in debug; C widened to int.
     let mut step = if diff >= 0 {
         diff >> shift
     } else {
-        -((-diff) >> shift)
+        -((-(diff as i32) >> shift) as i16)
     };
     if step == 0 {
         step = if diff > 0 { 1 } else { -1 };
